Improve Prospector Backup Directory Support
In newer FTB mods, backups are done with FTBUtilities, not AromaBackups.
FTBUtilities is set up to dump timestamped server zips directly into
<server>/backups without a <world_name> subdirectory. Using this
structure, prospector will never find backups.

This change adds a new option in the prospect config called
'per_world_backups', which is a boolean flag. If it's set to True, the
old AromaBackups style is used with <world_name> subdirectories within
the backups/ folder. If it's set to False, uses the new FTBUtilities
format, which skips the <world_name> subdirectories.

This hasn't been tested yet.