Instance Type Support
Added fields to both the Ansible and Terraform configurations. For
Terraform, allows easier management of the desired instance. For
Ansible, allows the playbook to do actions specific to the instance size
or category.

Ansible playbook now only runs the NVMe mounting task when the instance
category is "i3".

The README has been drastically expanded with some sample instance
costs, and mentions of the new "aws_instance_type" variables.